moved draw_target to a more appropriate place
